Split the Toolkit out of the IPC monolith
Six channel prefixes that make up one product surface — Convert, Apps, Translate,
Protect, Presenter and OCR Workspace — move together, 60 handlers in all:

  electron/ipc.ts     5148-5546 -> electron/ipc/toolkit.ts     (60 handlers)
  electron/preload.ts 1594-1722 -> electron/preload/toolkit.ts (68 bindings)
  shared/types.ts     9090-9205 -> shared/api/toolkit.ts       (65 signatures)

First domain with its own closure state, and it moves with the block rather than
into the shared context: toolkitSignals, translateSignals, presenterDir and
pendingPresenterImports are private to the toolkit now. toolkitCopy, the
native-dialog localizer, was declared 4,000 lines away from its only 25 callers
and is relocated next to them first, as a separate step.

Three toolkit tests read electron/ipc.ts and electron/preload.ts as text to check
things that are not about a channel at all — "the OCR manager is initialised with
the window", "deleting a generated document uses the OS Trash", "dropped files
become native paths only in preload". They now match against mainSourceText() /
bridgeSourceText() from the channel census, which concatenate the whole main and
bridge trees. The assertion keeps its meaning and stops depending on which file
the code sits in.

electron/ipc.ts is now 5,128 lines (from 6,013), electron/preload.ts 1,597 (from
1,862), shared/types.ts 9,214 (from 9,553).

Verified: 1,425 tests pass, lint and typecheck clean, and the IPC contract test
confirms every channel is still registered exactly once and reachable.
